+++ hardhat.config.ts
@@ -11,7 +11,7 @@ import './scripts/tasks/protocol/mintTalentLayerId'
 import './scripts/tasks/protocol/addArbitrator'
 import './scripts/tasks/protocol/removeArbitrator'
 import './scripts/tasks/protocol/updateMinArbitrationFeeTimeout'
-import './scripts/tasks/protocol/setTrustedForwarder'
+import './scripts/tasks/protocol/addTrustedForwarder'
 import { Network } from './scripts/config'
 
 dotenvConfig({ path: resolve(__dirname, './.env') })
