+++ scripts/tasks/protocol/addTrustedForwarder.ts
